requirements/local.txt
-r base.txt
pytest-django==3.2.1
factory-boy==2.11.1
